Support Vec<OpaqueRustType> as a fn arg
We also fix the `RustVec`'s `deinit` to only free the underlying
`Vec<T>` if the RustVec is owned. This prevents a double free
after passing an owned `RustVec` from Swift to Rust.
