+++ crates/swift-integration-tests/src/slice.rs
@@ -1,0 +1,21 @@
+// #[cfg(test)]
+// mod tests {
+//     use super::*;
+//
+//     #[test]
+//     fn todo() {
+//         unimplemented!(
+//             r#"
+// - Define SliceTestOpaqueRustType
+// - Add a method to create Vec<SliceTestOpaqueRustType>
+// - Add method to reflect &[OpaqueRustType]
+//   - Add test to verify that we can iterate over the slice
+//   - Add test to verify that that we can index into the slice
+//   - Add test to verify that we can get the length of the slice
+// - Create SliceTests.swift
+//   - Add Swift test verifying that we can use an Array<SliceTestOpaqueRustType> as a slice
+//   - Add Swift test verifying that we can use a RustVec<SliceTestOpaqueRustType> as a slice
+//         "#
+//         )
+//     }
+// }
