+++ crates/swift-integration-tests/src/vec.rs
@@ -10,10 +10,9 @@ mod ffi {
     }
 
     extern "Rust" {
-        fn create_vec_u8(contents: &[u8]) -> Vec<u8>;
-        fn create_vec_i32(contents: &[i32]) -> Vec<i32>;
-
-        fn create_vec_opaque_rust_type() -> Vec<ARustTypeInsideVecT>;
+        fn rust_reflect_vec_opaque_rust_type(
+            arg: Vec<ARustTypeInsideVecT>,
+        ) -> Vec<ARustTypeInsideVecT>;
     }
 }
 
@@ -33,16 +32,6 @@ impl ARustTypeInsideVecT {
     }
 }
 
-fn create_vec_u8(contents: &[u8]) -> Vec<u8> {
-    contents.to_vec()
-}
-
-fn create_vec_i32(contents: &[i32]) -> Vec<i32> {
-    contents.to_vec()
-}
-
-fn create_vec_opaque_rust_type() -> Vec<ARustTypeInsideVecT> {
-    vec![ARustTypeInsideVecT {
-        text: "hello there, friend".to_string(),
-    }]
+fn rust_reflect_vec_opaque_rust_type(arg: Vec<ARustTypeInsideVecT>) -> Vec<ARustTypeInsideVecT> {
+    arg
 }
